fix: Move files to correct locale, delete old analytics decorator
