+++ tests/fixtures/decorators/analytics_patches.py
